Completed handle_follow and moved SignatureVerification
